+++ docs/conf.py
@@ -226,7 +226,7 @@ html_static_path = ['_static']
 # Sphinx supports the following languages:
 #   'da', 'de', 'en', 'es', 'fi', 'fr', 'hu', 'it', 'ja'
 #   'nl', 'no', 'pt', 'ro', 'ru', 'sv', 'tr'
-#html_search_language = 'en'
+html_search_language = 'ko'
 
 # A dictionary with options for the search language support, empty by default.
 # Now only 'ja' uses this config value